+++ Changelog.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+* [#1202](https://github.com/mbj/mutant/pull/1202)
+
+  * Add `#reduce` -> `#sum` mutations
+    * `a.reduce(:+)`     -> `a.sum`
+    * `a.reduce(0, &:+)` -> `a.sum(0)`
+
 * [#1201](https://github.com/mbj/mutant/pull/1201)
 
   * Add `/\Astatic/` -> `#start_with?` mutations: